# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cleave API",
    "description": "Azure-first Cloud Security Posture Management.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/attack-paths": {
      "get": {
        "tags": [
          "attack-paths"
        ],
        "summary": "List Attack Paths",
        "description": "Every route from an exposed asset to a sensitive one, shortest first.\n\nBuilt from the assets and edges the last scan stored rather than from a\ntable of its own: a path is a pure function of those, and a stored one\ncould describe a route the customer has already closed.",
        "operationId": "list_attack_paths_api_v1_attack_paths_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Attack Paths Api V1 Attack Paths Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attack-paths/graph": {
      "get": {
        "tags": [
          "attack-paths"
        ],
        "summary": "Route Graph",
        "description": "Every route in the estate, as one graph, with what each link holds up.\n\nThe list above ranks routes, which is the right order for reading them and\nthe wrong one for acting: forty routes through one identity are forty rows\nthat never say \"one identity\". The drawing says it in one look, and the\nnumbers on its links say what closes if that identity's role goes -- for\nevery link, not for a shortlist, because they are all read off one analysis\n(``graph/severance.py``).\n\nOne request rather than three. The page used to ask for routes, then for\nchoke points, then for the outcome of a cut, and drew a shape in between\nthat it had to revise twice.",
        "operationId": "route_graph_api_v1_attack_paths_graph_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Route Graph Api V1 Attack Paths Graph Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attack-paths/choke-points": {
      "get": {
        "tags": [
          "attack-paths"
        ],
        "summary": "List Choke Points",
        "description": "The links worth cutting first, ranked by how many routes close with them.\n\nA separate endpoint rather than a field on the list above, because it costs\na re-traversal per candidate and the list is read far more often than the\nquestion is asked. A page that wants both asks for both.",
        "operationId": "list_choke_points_api_v1_attack_paths_choke_points_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "default": 5,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Choke Points Api V1 Attack Paths Choke Points Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attack-paths/blast-radius/{resource_id}": {
      "get": {
        "tags": [
          "attack-paths"
        ],
        "summary": "Blast Radius",
        "description": "What one identity or asset can act on.\n\nThe question a customer actually asks about an over-privileged principal:\nnever \"is this role too broad\" in the abstract, but \"what would go with it\".",
        "operationId": "blast_radius_api_v1_attack_paths_blast_radius__resource_id__get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Blast Radius Api V1 Attack Paths Blast Radius  Resource Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attack-paths/access/{resource_id}": {
      "get": {
        "tags": [
          "attack-paths"
        ],
        "summary": "Access",
        "description": "Who holds access to one asset, and what one identity holds.\n\nWhether or not anything exposed leads there -- the half of the question a\nroute cannot answer, because a route needs a way in. Read from the same\nper-role evaluation the routes are walked with, so a principal listed here\nas controlling an asset is exactly one a route may pass through\n(DECISIONS.md section 125).",
        "operationId": "access_api_v1_attack_paths_access__resource_id__get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Access Api V1 Attack Paths Access  Resource Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attack-paths/neighborhood/{resource_id}": {
      "get": {
        "tags": [
          "attack-paths"
        ],
        "summary": "Neighborhood",
        "description": "The assets around one, what reaches it and what it reaches.\n\nFor the graph view on an asset's page. The route list stays the answer to\n\"which link do I cut\"; this is for looking around (DECISIONS.md section 101).\nCapped at three hops: every route this graph can express is three or four\nhops end to end, and a neighbourhood that deep either way already spans it.",
        "operationId": "neighborhood_api_v1_attack_paths_neighborhood__resource_id__get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 1,
              "default": 2,
              "title": "Depth"
            }
          },
          {
            "name": "expand",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "maxItems": 20,
              "default": [],
              "title": "Expand"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Neighborhood Api V1 Attack Paths Neighborhood  Resource Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/attack-paths/simulate": {
      "post": {
        "tags": [
          "attack-paths"
        ],
        "summary": "Simulate",
        "description": "What closes if several links are removed together, across the organization.\n\nA POST because the plan is a body, not because anything is written: this\nchanges nothing, in CloudGuard or in the cloud, and a demo visitor may ask\nit like any other question. The answer is exact for the plan as a whole --\nnever the sum of what each link closes alone, which is the one number a\nplan of several changes cannot be read off (``AssetGraph.simulate``).",
        "operationId": "simulate_api_v1_attack_paths_simulate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Simulate Api V1 Attack Paths Simulate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SimulatedLink": {
        "properties": {
          "source": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Source"
          },
          "relationship": {
            "$ref": "#/components/schemas/RelationshipType"
          },
          "target": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Target"
          }
        },
        "type": "object",
        "required": [
          "source",
          "relationship",
          "target"
        ],
        "title": "SimulatedLink",
        "description": "One link, named the way a drawn edge names it."
      },
      "SimulationRequest": {
        "properties": {
          "cuts": {
            "items": {
              "$ref": "#/components/schemas/SimulatedLink"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Cuts"
          }
        },
        "type": "object",
        "required": [
          "cuts"
        ],
        "title": "SimulationRequest",
        "description": "Links to remove together. A body rather than a query string, because\nten Azure ids three times over do not fit in a URL every proxy accepts."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}